# Conflicts:
#	server/src/main/java/io/litmusblox/server/service/IJobCandidateMappingService.java
#	server/src/main/java/io/litmusblox/server/service/impl/JobCandidateMappingService.java

+++ server/src/main/java/io/litmusblox/server/service/JCMAllDetails.java
@@ -78,7 +78,7 @@ public class JCMAllDetails {
     Date hiredOn;
     String rejectedBy;
     Date rejectedOn;
-    Long offeredCompensation;
+    Double offeredCompensation;
     Date offerDetailsOfferedOn;
     Date joiningOn;
